refactor: Separate api and refactor logic use userId from principal
