+++ src/main/java/com/example/backend_webflux/handler/AuthHandler.java
@@ -1,11 +1,14 @@
 package com.example.backend_webflux.handler;
 
 import com.example.backend_webflux.domain.Token;
+import com.example.backend_webflux.domain.User;
 import com.example.backend_webflux.dto.AuthDto.Request;
 import com.example.backend_webflux.service.SecurityService;
+import com.example.backend_webflux.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.reactive.function.server.ServerRequest;
@@ -19,6 +22,7 @@ import reactor.core.publisher.Mono;
 public class AuthHandler {
 
   private final SecurityService securityService;
+  private final UserService userService;
 
   public Mono<ServerResponse> login(ServerRequest request) {
     Mono<Request> dto = request.bodyToMono(Request.class);
@@ -31,5 +35,14 @@ public class AuthHandler {
           return t;
         }));
     return ServerResponse.ok().body(response, Token.class);
+  }
+  public Mono<ServerResponse> modifyPw(ServerRequest request) {
+
+    Mono<User> updatedUser = Mono.zip(request.bodyToMono(User.class), request.principal())
+        .flatMap(z -> userService.modifyPw(z.getT2().getName(), z.getT1()));
+
+    return ServerResponse.ok()
+        .contentType(MediaType.APPLICATION_JSON)
+        .body(updatedUser, User.class);
   }
 }
