+++ src/main/java/com/example/backend_webflux/handler/UserHandler.java
@@ -8,7 +8,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Component
@@ -17,20 +16,12 @@ public class UserHandler {
 
   private final UserService userService;
 
-  public Mono<ServerResponse> getAllUser(ServerRequest request) {
-    Flux<User> users = userService.getAllUser();
-
-    return ServerResponse.ok()
-        .contentType(MediaType.APPLICATION_JSON)
-        .body(users, User.class);
-  }
 
   @PreAuthorize("hasRole('USER')")
   public Mono<ServerResponse> getUser(ServerRequest request) {
 
     Mono<User> user = request.principal()
         .flatMap(auth -> userService.getUserById(auth.getName()));
-
 
     return ServerResponse.ok()
         .contentType(MediaType.APPLICATION_JSON)
@@ -48,10 +39,9 @@ public class UserHandler {
   }
 
   public Mono<ServerResponse> modifyUser(ServerRequest request) {
-    Mono<User> unsavedUser = request.bodyToMono(User.class);
-    String id = request.pathVariable("id");
 
-    Mono<User> updatedUser = unsavedUser.flatMap(u -> userService.modify(id, u));
+    Mono<User> updatedUser = Mono.zip(request.bodyToMono(User.class), request.principal())
+        .flatMap(z -> userService.modify(z.getT2().getName(), z.getT1()));
 
     return ServerResponse.ok()
         .contentType(MediaType.APPLICATION_JSON)
@@ -59,20 +49,12 @@ public class UserHandler {
   }
 
   public Mono<ServerResponse> deleteUser(ServerRequest request) {
-    String id = request.pathVariable("id");
-    Mono<Void> deleted = userService.delete(id);
+    Mono<Void> deleted = request.principal()
+        .flatMap(auth -> userService.delete(auth.getName()));
 
     return ServerResponse.ok()
         .contentType(MediaType.APPLICATION_JSON)
         .body(deleted, Void.class);
   }
 
-  public Mono<ServerResponse> modifyPw(ServerRequest request) {
-
-
-
-    return ServerResponse.ok()
-        .contentType(MediaType.APPLICATION_JSON)
-        .body(null, User.class);
-  }
 }
